hotfix

+++ backend/internal/models/resp.go
@@ -35,7 +35,7 @@ type PaymentDTO struct {
 	Currency     string  `json:"currency"`
 	Provider     string  `json:"provider"`
 	Amount       int     `json:"amount"`
-	PaymentDt    int64   `json:"payment_dt"`
+	PaymentDt    int     `json:"payment_dt"`
 	Bank         string  `json:"bank"`
 	DeliveryCost float64 `json:"delivery_cost"`
 	GoodsTotal   float64 `json:"goods_total"`
